skills/pyproject.toml
[project]
name = "skills"
version = "0.1.0"
description = "Shared skills layer — reusable tools and playbooks across all agents"
requires-python = ">=3.10"
dependencies = [
    "shared",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "langchain-core>=0.3",
    "pdfplumber>=0.11",
    "pytesseract>=0.3",
    "pillow>=10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "reportlab>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skills"]
